backtest/requirements.txt
Click==7.0
Flask==1.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
Werkzeug==0.14.1
